--- api/src/main/kotlin/com/pokitmonz/pokit/PokitApplication.kt
@@ -1,11 +1,0 @@
-package com.pokitmonz.com.pokitmonz.pokit
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication(scanBasePackages = ["com.pokitmonz"])
-class PokitApplication
-
-fun main(args: Array<String>) {
-    runApplication<PokitApplication>(*args)
-}
