+++ entry/web/src/main/kotlin/com/pokit/web/WebApplication.kt
@@ -1,0 +1,11 @@
+package com.pokit.web
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class WebApplication
+
+fun main(args: Array<String>) {
+    runApplication<WebApplication>(*args)
+}
