--- infra/src/test/kotlin/com/pokitmonz/pokit/InfraTestApplication.kt
@@ -1,11 +1,0 @@
-package com.pokitmonz.pokit
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class InfraTestApplication
-
-fun main(args: Array<String>) {
-    runApplication<InfraTestApplication>(*args)
-}
