--- infra/src/test/kotlin/com/pokitmonz/pokit/common/config/TestAuditingConfig.kt
@@ -1,8 +1,0 @@
-package com.pokitmonz.pokit.common.config
-
-import org.springframework.boot.test.context.TestConfiguration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@TestConfiguration
-@EnableJpaAuditing
-class TestAuditingConfig
