--- infra/src/test/kotlin/com/pokitmonz/pokit/example/InfraContainerTest.kt
@@ -1,11 +1,0 @@
-package com.pokitmonz.pokit.example
-
-import com.pokitmonz.pokit.common.support.DataJpaTestSupport
-import org.junit.jupiter.api.Test
-
-class InfraContainerTest : DataJpaTestSupport() {
-    @Test
-    fun test() {
-        println("hello")
-    }
-}
