+++ settings.gradle.kts
@@ -1,5 +1,9 @@
 rootProject.name = "pokit"
 include("domain")
-include("core")
-include("api")
-include("infra")
+include("application")
+include("entry")
+include("entry:web")
+include("adapters")
+include("adapters:in-web")
+include("adapters:out-persistence")
+include("adapters:out-cache")
